pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seamless-transformer"
version = "0.3.0"
description = "Computation engine for Seamless: execute checksum-addressed transformations in Python or bash"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Sjoerd de Vries"}]
dependencies = [
    "bashlex",
    "PyYAML",
    "tqdm",
    "seamless-core",
]

[project.optional-dependencies]
compiled = ["cffi", "numpy", "seamless-signature"]

keywords = ["seamless", "transformer", "computation", "checksum", "reproducible"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://github.com/sjdv1982/seamless"
Repository = "https://github.com/sjdv1982/seamless-transformer"
Issues = "https://github.com/sjdv1982/seamless/issues"

[project.scripts]
seamless-run = "seamless_transformer.cmd.api.main:main"
seamless-upload = "seamless_transformer.cmd.api.upload:main"
seamless-download = "seamless_transformer.cmd.api.download:main"
seamless-run-transformation = "seamless_transformer.api.run_transformation:main"
seamless-queue = "seamless_transformer.api.queue:main"
seamless-queue-finish = "seamless_transformer.api.queue:finish"

[tool.setuptools]
script-files = ["bin/_seamlessify", "bin/seamless-mode-bind.sh"]

[tool.setuptools.packages.find]
where = ["."]
include = ["seamless_transformer*"]
exclude = ["tests*"]
